Added an observer to the notification permission status
